requirements.txt
gradio
mcp
pandas
python-dotenv